Register application works.  Add preliminary INSTALL.md, rename io.py.
As mentioned in [this python bug](https://bugs.python.org/issue21202), naming a
file io.py can cause a name collision, and it did for me here. I renamed io.py
to formerly_io.py and changed references to it in dialogs.py to call formerly_io
instead of io.  I suspect there will be some other places where we need to make
this change (io -> some_other_name).

This commit marks the first time that the register application has worked on my
local installation.  I installed the latest version of SqlAlchemy (1.0.4)
(instead of making the #5 changes) using pip, as documented in INSTALL.md.  As
far as I can tell, the app is working properly with that version of SqlAlchemy,
but we'll need to draw up a full set of tests to make sure that it behaves as
expected (see issue #10).