Start sample data.
Add a file that loads the inventory tables with one line of data in each.  Some
questions: the inventory clerks table does not correspond to the clerks
available in the register version.  Should we drop it?  Similarly, the
inventory.coupon_items table doesn't seem to appear in the code, nor does
register_tape.test_table.